# use atomic.Pointer[T] instead unsafe pointers