--- a/.github/workflows/build.yml
+++ b/.github/workflows/build.yml
@@ -22,7 +22,7 @@ jobs:
     - name: Setup Go
       uses: actions/setup-go@v2
       with:
-        go-version: '1.17.0' # The Go version to download (if necessary) and use.
+        go-version: '1.19.0' # The Go version to download (if necessary) and use.
 
     # Install all the dependencies
     - name: Install dependencies