as I mention on https://github.com/leporo/sqlf/pull/14 this PR uses the new atomic api available since go 1.19 instead use unsafe pointers.

also, this is a big improvement since the original code uses atomic operations to write to the pointer, but the code does not uses atomic operations to read, so it is an optimistic approach but it is not really thread safe. Now each operation use atomics and I think the performance impact is minimal (if not, we can wrap the pointer in a structure that uses a RW mutex). I run the benchmarks before and after, you can see the results below

unfortunately this requires go 1.19 and this may be not a problem since the latest stable go version at this moment is go 1.23.4

if you decided to merge https://github.com/leporo/sqlf/pull/13 first, then the go version matrix should be updated to not run tests on older go versions

```console
oos: linux
goarch: amd64
pkg: github.com/leporo/sqlf
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                              │   old.txt    │               new.txt                │
                              │    sec/op    │    sec/op     vs base                │
SelectDontClose-8               1.236µ ± ∞ ¹   1.607µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Select-8                        149.7n ± ∞ ¹   153.9n ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectPg-8                      150.1n ± ∞ ¹   151.8n ± ∞ ¹       ~ (p=1.000 n=1) ²
ManyFields-8                    1.958µ ± ∞ ¹   1.843µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Bind-8                          637.9n ± ∞ ¹   668.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
ManyFieldsPg-8                  1.974µ ± ∞ ¹   2.128µ ± ∞ ¹       ~ (p=1.000 n=1) ²
MixedOrder-8                    156.1n ± ∞ ¹   171.0n ± ∞ ¹       ~ (p=1.000 n=1) ²
BuildPg-8                       24.96n ± ∞ ¹   25.13n ± ∞ ¹       ~ (p=1.000 n=1) ²
Build-8                         26.89n ± ∞ ¹   25.53n ± ∞ ¹       ~ (p=1.000 n=1) ²
Dest-8                          111.4n ± ∞ ¹   115.6n ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectComplex-8                 527.6n ± ∞ ¹   517.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectComplexPg-8               532.7n ± ∞ ¹   612.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryFmt-8             1.141µ ± ∞ ¹   1.140µ ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryFmtPostgreSQL-8   1.107µ ± ∞ ¹   1.183µ ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubquery-8                488.9n ± ∞ ¹   480.7n ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryPostgreSQL-8      484.7n ± ∞ ¹   630.1n ± ∞ ¹       ~ (p=1.000 n=1) ²
With-8                          653.4n ± ∞ ¹   644.7n ± ∞ ¹       ~ (p=1.000 n=1) ²
In-8                            487.3n ± ∞ ¹   502.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                         374.9n         394.0n        +5.08%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                              │   old.txt   │               new.txt               │
                              │    B/op     │    B/op      vs base                │
SelectDontClose-8               624.0 ± ∞ ¹   624.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select-8                        0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectPg-8                      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
ManyFields-8                    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Bind-8                          24.00 ± ∞ ¹   24.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
ManyFieldsPg-8                  0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
MixedOrder-8                    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
BuildPg-8                       0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Build-8                         0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Dest-8                          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectComplex-8                 0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectComplexPg-8               0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryFmt-8             112.0 ± ∞ ¹   112.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryFmtPostgreSQL-8   112.0 ± ∞ ¹   112.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubquery-8                0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryPostgreSQL-8      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
With-8                          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
In-8                            0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                   ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean

                              │   old.txt   │               new.txt               │
                              │  allocs/op  │  allocs/op   vs base                │
SelectDontClose-8               9.000 ± ∞ ¹   9.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select-8                        0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectPg-8                      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
ManyFields-8                    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Bind-8                          3.000 ± ∞ ¹   3.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
ManyFieldsPg-8                  0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
MixedOrder-8                    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
BuildPg-8                       0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Build-8                         0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Dest-8                          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectComplex-8                 0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectComplexPg-8               0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryFmt-8             2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryFmtPostgreSQL-8   2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubquery-8                0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
SelectSubqueryPostgreSQL-8      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
With-8                          0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
In-8                            0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                   ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean
```