@peczenyj 2025-01-11
BTW this approach seems overkill, most of the implementations just wa that change a singleton value is not thread safe