@leporo 2025-02-24
Thank you for spotting this. 

I can't remember why I used atomic assignment for default dialect. I believe it had to be selected long before any parallel processing starts, so there is no need for thread safety. If an application uses multiple dialects simultaneously, it probably shouldn't use the default one and do `dialect.From("table")...` instead of `sqlf.From`.

I would rather remove the atomic assignment than use the atomic read of default dialect. As there is a minor chance the atomic assignment is there for some reason (which I can't remember), let's keep it as is.